feat: rewrite audit script in Python.
The new Python script is a standalone executable in
/usr/bin/audit-secureblue. It currently carries out the same checks as
the old bash audit script, but is faster and more modular.

Each check is a generator that accepts some state (an associative array
that can be used to pass information between checks) and yields a
sequence of reports and recommendations, which are handled by the audit
runner.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>