feat: update run0edit to v0.4.0 w/ support for immutable flag
The new version of `run0edit` checks for the immutable flag on the file
to be edited, and if so, asks the user whether to temporarily remove it
to edit the file. This makes editing immutable files less error-prone,
as the immutable flag is only removed for the brief window necessary to
overwrite the file contents and reapplied immediately afterward; the
script also compares the file contents after the immutable flag has been
reapplied to ensure it was not modified by another program during that
window.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>