feat(audit): merge flatpak recommendations by permission
This makes the recommendations associated to flatpak permissions group
the recommendations by permission, giving a list of flatpaks that have
each flagged permission. This significantly shortens the recommendation
list in many cases, resulting in the recommendations being more usable
as users can see at a glance which flatpaks have a given permission.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>