feat: enhance bash env lockdown by preventing moving .config
This enhances the bash environment lockdown by changing certain
directory permissions to prevent unprivileged users from moving or
deleting the subdirectories .config, .cache, and .local of their
home directory, while still allowing users full read-write access to
those directories.

This is accomplished as follows:
1. Change the modes of the user's home directory and the above
   subdirectories to 1770. The sticky bit prevents the user from moving
   or deleting contained files or directories they do not own.
2. Use `setfacl` to grant the user rwx access to the directories. (This
   avoids problems if one of these directories has an ACL set that
   overrides the default group access permissions.)
3. Change the directories' ownership to `root:user`. Because of the
   first two steps, the user still has full access to the directories'
   contents.

This is admittedly somewhat of a hack, but it works and shouldn't
restrict user access to their files in any unintended ways.

Also added a check to the audit script for these permission settings.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>