fix: set XDG_CONFIG_HOME explicitly
This avoids podman checking whether `~/.config` is owned by the current
user and erroring out if it doesn't. Unclear how long this can be relied
on but it works for now and in any case is harmless since we're just
setting these to their defaults.

Also set the other xdg base directories to their default values for
consistency.