feat: make default bash prompt display nonzero exit codes
If the previous command's exit code was nonzero, the exit code is
displayed inside square brackets in bold and red at the beginning of the
bash command prompt. This gives a visual indicator of when a process has
silently failed.

The file needs to be alphabetically later than vte.sh, since that script
in /etc/profile.d overwrites PROMPT_COMMAND.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>