fix: check bash session is interactive before modifying prompt
Some scripts check whether the prompt variable `PS1` is set to determine
whether the shell is interactive. While this is less robust than
checking `$-`, it's still a method mentioned in the bash manual, so we
should make sure not to break this behavior and only modify `PS1` if the
shell is interactive.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>