chore: Remove regenerate-grub from utilities.just
Due to upstream changes in Fedora Atomic, grub configuration is now static, not dynamic. It is now expected that the only changes made to the main grub.cfg file come from upstream, not from manual user configuration. The static config supports sourcing user config from a file in the same directory named user.cfg. 

I cannot tell if the above is related to the composefs path issue, but either way this script is no longer needed in secureblue.