chore: Move KDE Plasma service overrides to /usr
These overrides just add an `ExecCondition` that prevents the services
from running if Xwayland is disabled. These services fail anyway with
Xwayland disabled. Since the overrides are set by secureblue rather than
by the local admin, it makes sense for them to be in `/usr` instead of
`/etc`.

(Split from #1447.)

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>