feat: add service to update SELinux contexts in /var
Changes to secureblue's fork of SELinux policy will require changes to
file contexts in `/var`, which won't be applied automatically on update.
This is especially an issue for the flatpak policy being developed. To
handle this, we set up a service to run `restorecon` on the affected
files, with a version number to allow for re-running as needed with
future updates.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>