fix: remove includepkgs line from VPN repos (#1562)
Proton recently split up the package for the VPN, resulting in a new
package not included in the includepkgs list. I'm not sure why such a
list was added in the first place; the VPN repos are disabled unless the
user chooses to enable them, at which point we might as well allow the
user to install any package from the repo.