feat(selinux): make Trivalent policy role-agnostic
Currently the Trivalent SELinux policy implicitly assumes that the user
is unconfined, and doesn't work for confined users. This modifies the
policy to work for confined users as well.

The `trivalent_t` and `trivalent_script_t` domains are replaced by
`trivalent_domain` and `trivalent_script_domain` attributes, containing
domains `$1_trivalent_t` and `$1_trivalent_script_t`, respectively, for
each role `$1_r` permitted to run Trivalent.

The use of domain prefixes allows for role-dependent domain transitions,
so for example `staff_trivalent_t` can be authorized to run something as
`staff_$2_t`. This will be useful once the new flatpak policy is in use,
and I included commented-out lines that will allow for Trivalent-flatpak
integration to keep working after we switch to that policy.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>