feat(audit): check user group memberships
Membership in groups like `docker` or `libvirt` grants the user access
equivalent to root. People sometimes add their main user account to
these groups for convenience, possibly without being aware of the
security implications. This adds a new check to the audit script that
flags membership in these groups and informs the user.

Additionally, the `systemd-journal` group is flagged as a warning due to
allowing viewing the kernel logs, and other unrecognized groups are
flagged with a generic warning about the potential security risks of
group memberships.

The only groups that aren't flagged are the user's username group,
`wheel` (which is already checked earlier), and `usbguard`.