feat: rewrite ujust harden-flatpak in Python
The Python rewrite has the same basic logic but with some improvements
to user-facing messages and the app ID search logic.

The third-party `inquirer` module is used for letting the user select
from a list of options, and is installed via RPM. (I put it in
`common-packages.yml` rather than `desktop-packages.yml` because it's
likely to be useful in other ujust scripts, and is a small package in
any case.)

Also move audit-script-specific parts of the `utils` module into a new
`audit_utils` module, and fix a minor type error in `is_using_vpn`.