chore: use manuallyinstalljust.sh in tests
This ensures the version of `just` used in CI is the same as the one
used in the images. Also made some minor reliability improvements to
`manuallyinstalljust.sh`: use `mktemp` to generate a temporary directory
instead of just using `/tmp`, and set `curl` options to retry on failure
and to not suppress error message output.