feat: verify provenance before automatic updates
Add an `ExecStartPre` item to `rpm-ostreed-automatic.service` that
verifies build provenance for the image to be pulled.

Technically this is subject to a TOCTOU issue since a new image could be
built after provenance verification is run and before the update itself
starts. However, this is a very small window of time, and it would be
difficult to avoid this without integrating provenance verification into
`rpm-ostree` itself.