refactor(install-trivalent.sh): dnf4 -> dnf5 preserving GPG-checking
Previously, DNF4 was used instead of DNF5 for downloading the Trivalent package, as DNF4 performs GPG check after download and DNF5 does not. Neither DNF4 nor DNF5 check the signature of a package installed from outside of a repository (by local path or URL) by default. This commit switches to using DNF5 to download the package (without checking the signature at that point), and subsequently forcing a signature check when installing this package from a local file.