fix(audit): make strings coming from argparse show up in POT file
A few strings that are printed as part of the help text are marked as
translatable inside the `argparse` module, but they haven't been getting
included in the POT file, because they're not present in the audit
script's source code, which is what xgettext scans to generate the POT
file.

We can fix this by simply including those same strings in an unused list
in the script.